requirements.txt
ipykernel
pillow
numpy
torch
torchvision
matplotlib
tqdm
scikit-learn
brevitas